+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 0.21
 
 * #26 Remove API deprecated since 0.15
+* #27 Upgrade to bootique 0.22 and bootique-jetty 0.20
 
 ## 0.20